Add configuration file which defines annotation groups
Adding this file to a Girder/DSA instance removes the need for manual specification
of annotation groups in HistomicsUI. Easy switching between annotation groups
is desirable for quick correction of AI miss-classifications.

The file is to be put in the collection level, or folder in which predictions will be uploaded.

For further description, see:
https://github.com/DigitalSlideArchive/HistomicsUI/blob/master/docs/config_options.rst